fix/int to float

+++ lb_script/utils.py
@@ -9,7 +9,7 @@ def read_yaml(filepath):
     return data
 
 def check_int(value):
-    return 0 if math.isnan(value) else int(value)
+    return 0 if math.isnan(value) else float(value)
 
 def check_str(value):
     return '' if math.isnan(value) else str(value)
